+++ Src/Library/changelog.md
@@ -1,2 +1,5 @@
 ### EXPERIMENTAL
 - support for `Asp.Versioning.Http` package via `FastEndpoints.AspVersioning` #352
+
+### IMPROVEMENTS
+- nswag operation processor internal optimizations